tests for #491/492, webp lossless abi
